+++ src/Strain.cpp
@@ -6,12 +6,12 @@
 using namespace std;
 using namespace zppsim;
 
-Strain::Strain(size_t id, std::vector<GenePtr> const & genes) :
+Strain::Strain(int64_t id, std::vector<GenePtr> const & genes) :
 	id(id), genes(genes)
 {
 }
 
-size_t Strain::size()
+int64_t Strain::size()
 {
 	return genes.size();
 }
@@ -21,7 +21,7 @@ std::vector<GenePtr> Strain::getGenes()
 	return genes;
 }
 
-GenePtr Strain::getGene(size_t index)
+GenePtr Strain::getGene(int64_t index)
 {
 	return genes[index];
 }
